Instrumentation Shutoff was not designed with multiple images in mind. This commit extends the original dynamic instrumentation code to include an image sequence number in the key (if desired). Since the image sequence number is not known until runtime, the key is initialized by the tool with an image sequence of 0, and then is changed at runtime initialization.
Since the runtime initialization cannot tell the difference between an
initialization dynamic point (e.g. tool_image_init) that uses the image hash
as the key and a user-defined dynamic point (e.g. function_entry) that uses
a combination of the function sequence number, image sequence number, and
dynamic point type as the key, we had to also give initialization dynamic
types a type: PointType_inits.

This commit modifies:

* the pebil source code (ElfFileInst and InstrumentationTool),
and AddressStreamIntercept, BasicBlockCounter, FunctionTimer, and
LoopIntercept tools to label initialization dynamic points with the
PointType_inits point type.
* the Metasim file to include the PointType_inits point type and
functionality for generating a unique key given a sequence id and an image
sequence id.
* the DynamicInstrumentation class to initialize dynamic points with a
given image sequence number (unless it is of type PointType_inits) and
include a print function based on the key.
* the AddressStreamIntercept tool to use the image sequence as a part
of the key for memory op instrumentation points.
* the FunctionTimer tool to use the image sequence as a part of the key
for function entry/exit instrumentation points.
* the AddressStreamDriver, CounterFunctions, LoopTimer, PAPIFunc, PAPIInst,
and TimerFunctions libraries to recognize the new initialization dynamic
point keys.
* the AddressStreamDriver, LoopTimer, PAPIFunc, PAPIInst, and
TimerFunctions libraries to only allocate the DynamicInstrumentation
object once (and not per-image) and to only finalize images once (not
per-image).
* the AddressStreamDriver to use 64-bit types for the memory sequence
number and use the new image sequence number in dynamic point keys to
determine which image (and therefore which group) the memory op belongs
to. Highly important for shutting off instrumentation for a group.

+++ instcode/AddressStreamDriver.hpp
@@ -118,8 +118,8 @@ class AddressStreamDriver {
     virtual void SetUpTools();
 
     void ShutOffInstrumentationInAllBlocks();
-    void ShutOffInstrumentationInBlock(uint32_t blockID);
-    void ShutOffInstrumentationInBlocks(std::set<uint32_t>& blocks);
+    void ShutOffInstrumentationInBlock(uint64_t blockID);
+    void ShutOffInstrumentationInBlocks(std::set<uint64_t>& blocks);
     void ShutOffInstrumentationInMaxedGroups(image_key_t, thread_key_t);
 
     // For Testing Purposes
